// frontend/src/types/project.ts
export interface Project {
  id: number;
  name: string;
  description?: Record<string, unknown> | null;
  company_id: number;
  executive?: string | null;
  executive_email?: string | null;
  city?: string | null;
  region?: string | null;
  address?: string | null;
  street_number?: string | null;
  commune?: string | null;
  country?: string | null;
  postal_code?: string | null;
  url_google_maps?: string | null;
  latitude?: string | null;
  longitude?: string | null;
  phone?: string | null;
  website?: string | null;
  price_from?: number | null;
  price_to?: number | null;
  unique_price?: number | null;
  currency: string;
  total_units?: number | null;
  available_units?: number | null;
  type?: string | null;
  delivery_type?: string | null;
  state: string;
  is_active: boolean;
  is_published: boolean;
  is_promo: boolean;
  is_hot: boolean;
  start_date?: string | null;
  delivery_date?: string | null;
  delivery_end_date?: string | null;
  launch_date?: string | null;
  after_sales_email?: string | null;
  after_sales_executive?: string | null;
  after_sales_phone?: string | null;
  finances_insurance_company?: string | null;
  finances_insurance_policy?: string | null;
  finances_insured_value?: string | null;
  finances_insured_value_util?: string | null;
  finances_benefit?: string | null;
  finances_annual_rate?: number | null;
  finances_credit_rate?: number | null;
  down_payment_bonus?: number | null;
  reservation?: number | null;
  pre_approval?: string | null;
  installments?: number | null;
  balloon_payment?: string | null;
  annual_capital_gain?: number | null;
  construction_capital_gain?: number | null;
  vacancy?: number | null;
  additional_information?: string | null;
  initial_payment?: number | null;
  big_down_payment?: number | null;
  notes_rent?: Record<string, unknown> | null;
  social_networks?: Record<string, unknown> | null;
  payment_plan_conditions?: Record<string, unknown> | null;
  promotions_benefits?: Record<string, unknown> | null;
  frequently_asked_questions?: Record<string, unknown> | null;
  created_at: string;
  updated_at: string;
}